requirements.txt
uvicorn
fastapi
icalendar
pytz
requests
mysql.connector